Begin work on migration to new Spring Data & Mongo
Initial work to transition to Spring Data MongoDB 3.x and Mongo 4.x
driver. There are a lot of TODOs, and in order for the tests to work,
I had to copy the MongoServerExtension from kiwi-test and modify
it to work with the Mongo 4.x driver.

We also discussed removing support for Mongo 3.x driver and
removing the "automatic discovery" code and instead requiring
users (i.e. us) to pass in the ConnectionDriver explicitly.